feat: enforce task ownership in TaskService and update related methods in TaskController

+++ src/main/java/dev/guilhermeluan/todo_list/service/TaskService.java
@@ -6,6 +6,7 @@ import dev.guilhermeluan.todo_list.exceptions.NotFoundException;
 import dev.guilhermeluan.todo_list.model.Priority;
 import dev.guilhermeluan.todo_list.model.Task;
 import dev.guilhermeluan.todo_list.model.TaskStatus;
+import dev.guilhermeluan.todo_list.model.User;
 import dev.guilhermeluan.todo_list.repository.TaskRepository;
 import dev.guilhermeluan.todo_list.repository.TaskSpecification;
 import org.springframework.data.domain.Page;
@@ -18,9 +19,11 @@ import java.time.LocalDate;
 @Service
 public class TaskService {
     private final TaskRepository repository;
+    private final UserService userService;
 
-    public TaskService(TaskRepository repository) {
+    public TaskService(TaskRepository repository, UserService userService) {
         this.repository = repository;
+        this.userService = userService;
     }
 
     public Page<Task> findAll(Long userId, TaskStatus status, Priority priority, LocalDate dueDate, Pageable pageable) {
@@ -37,13 +40,17 @@ public class TaskService {
         return repository.save(task);
     }
 
-    public void update(Task taskToUpdate){
+    public void update(Task taskToUpdate, Long userId) {
         Task taskFound = findByIdOrThrowNotFound(taskToUpdate.getId());
+        User user = userService.findUserByIdOrThrowNotFound(userId);
+
+        validateTaskOwnership(taskFound, userId);
 
         if (taskToUpdate.getStatus() == TaskStatus.DONE && !taskToUpdate.isSubTask()) {
             assertThatAllSubTasksAreCompleted(taskFound);
         }
 
+        taskToUpdate.setUser(user);
         taskToUpdate.setSubTasks(taskFound.getSubTasks());
 
         repository.save(taskToUpdate);
@@ -52,9 +59,7 @@ public class TaskService {
     public Task createSubTask(Long parentId, Task subTask, Long userId) {
         Task parentTask = findByIdOrThrowNotFound(parentId);
 
-        if (!parentTask.getUser().getId().equals(userId)) {
-            throw new ForbiddenException("A tarefa pai não pertence ao usuário autenticado.");
-        }
+        validateTaskOwnership(parentTask, userId);
 
         if (parentTask.isSubTask()) {
             throw new BadRequestException("Não é possível aninhar subtarefas. A tarefa pai deve ser uma tarefa principal");
@@ -66,8 +71,9 @@ public class TaskService {
         return repository.save(subTask);
     }
 
-    public void delete(Long id) {
-        assertTaskExists(id);
+    public void delete(Long id, Long userId) {
+        Task task = findByIdOrThrowNotFound(id);
+        validateTaskOwnership(task, userId);
         repository.deleteById(id);
     }
 
@@ -75,13 +81,17 @@ public class TaskService {
         findByIdOrThrowNotFound(id);
     }
 
-    public Task updateStatus(TaskStatus newStatus, Long id) {
+    public Task updateStatus(TaskStatus newStatus, Long id, Long userId) {
         Task existingTask = findByIdOrThrowNotFound(id);
+        User user = userService.findUserByIdOrThrowNotFound(userId);
+
+        validateTaskOwnership(existingTask, userId);
 
         if (newStatus == TaskStatus.DONE && !existingTask.isSubTask()) {
             assertThatAllSubTasksAreCompleted(existingTask);
         }
 
+        existingTask.setUser(user);
         existingTask.setStatus(newStatus);
         return repository.save(existingTask);
     }
@@ -92,6 +102,12 @@ public class TaskService {
 
         if (hasIncompleteSubTasks) {
             throw new BadRequestException("Conclua todas as subtarefas pendentes antes de finalizar a tarefa principal.");
+        }
+    }
+
+    private void validateTaskOwnership(Task task, Long userId) {
+        if (!task.getUser().getId().equals(userId)) {
+            throw new ForbiddenException("A tarefa não pertence ao usuário autenticado.");
         }
     }
 }
